applies config only if target volume is mounted

+++ habitat-config/update-config.sh
@@ -3,13 +3,15 @@
 cd "/habitat-config" || exit 1
 
 for dirPath in "./source/"*; do
-    echo "Applying configuration for '$APP_MODULE_NAME' ($(basename "$dirPath"))"
+    echo "*** Applying configuration for '$APP_MODULE_NAME' (Source directory: $(basename "$dirPath"))"
     # Traefik Dynamic Config
-    echo "Applying traefik config"
-    if [ -d "./source/$dirPath/traefik" ]; then
-        rm -f "./target/traefik/${APP_MODULE_NAME}."*
-        for f in "./source/$dirPath/traefik/"*; do
-            cp "$f" "./target/traefik/${APP_MODULE_NAME}.$(basename "$f")";
-        done
+    if [ -d "./target/traefik" ]; then
+        echo "Applying traefik config"
+        if [ -d "./source/$dirPath/traefik" ]; then
+            rm -f "./target/traefik/${APP_MODULE_NAME}."*
+            for f in "./source/$dirPath/traefik/"*; do
+                cp "$f" "./target/traefik/${APP_MODULE_NAME}.$(basename "$f")";
+            done
+        fi
     fi
 done
